requirements.txt
click==8.1.8
colorama==0.4.6
joblib==1.4.2
nltk==3.9.1
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2025.2
xgboost==2.1.4
